Update user.rs
Converted user.rs into user/ directory-style module.

Implemented FromGodot, GodotConvert, and ToGodot on User; Required implementing all three traits on every other data type contained by User.

Converted all usize members to i64, to more easily convert back and forth between Godot's Variant type.

+++ rust/src/nodes/steamachievement.rs
@@ -19,7 +19,7 @@ pub struct SteamAchievementNode
 	#[base]
 	base: Base<HBoxContainer>,
 	
-	pub appId: u32,
+	pub appId: i64,
 	pub achievement: SteamAchievement,
 	
 	#[export]
