+++ rust/src/platforms/steam/data/playerachievements.rs
@@ -16,8 +16,8 @@ pub struct GetPlayerAchievementsPayload
 pub struct PlayerAchievement
 {
 	pub apiname: String,
-	pub achieved: usize,
-	pub unlocktime: usize,
+	pub achieved: i64,
+	pub unlocktime: i64,
 	pub name: String,
 	pub description: Option<String>,
 }
